backend/requirements.txt
# Core
fastapi==0.121.3
uvicorn==0.38.0
python-dotenv==1.2.1
pydantic==2.12.4

# AI
openai==2.8.1

# Auth
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# Database
aiosqlite==0.21.0
databases==0.9.0

# HTTP
httpx==0.28.1
aiofiles==24.1.0

# CORS & middleware
starlette==0.50.0

# Utilities
PyYAML==6.0.3
colorama==0.4.6
tqdm==4.67.1
